Type information for DuplicateNotAllowedException (similar to NotFoundException)
Helper super class TypedApiException that is inherited by NotFoundException and DuplicateNotAllowedException
