+++ src/java/com/unifina/api/DuplicateNotAllowedException.java
@@ -1,7 +1,15 @@
 package com.unifina.api;
 
-public class DuplicateNotAllowedException extends ApiException {
+public class DuplicateNotAllowedException extends TypedApiException {
+	public DuplicateNotAllowedException(String message, String type, String id) {
+		super(400, "DUPLICATE_NOT_ALLOWED", message, type, id);
+	}
+
+	public DuplicateNotAllowedException(String type, String id) {
+		this(type + " with id " + id + " already exists", type, id);
+	}
+
 	public DuplicateNotAllowedException(String message) {
-		super(400, "DUPLICATE_NOT_ALLOWED", message);
+		this(message, null, null);
 	}
 }
