+++ src/java/com/unifina/api/NotFoundException.java
@@ -1,31 +1,15 @@
 package com.unifina.api;
 
-public class NotFoundException extends ApiException {
-	/** Type of the missing resource */
-	private String type;
-	/** ID of the missing resource */
-	private String id;
-
+public class NotFoundException extends TypedApiException {
 	public NotFoundException(String message, String type, String id) {
-		super(404, "NOT_FOUND", message);
-		this.type = type;
-		this.id = id;
+		super(404, "NOT_FOUND", message, type, id);
 	}
+
 	public NotFoundException(String type, String id) {
 		this(type + " with id " + id + " not found", type, id);
 	}
+
 	public NotFoundException(String message) {
 		this(message, null, null);
-	}
-
-	@Override
-	public ApiError asApiError() {
-		ApiError e = super.asApiError();
-		if (type != null && id != null) {
-			e.addToBody("type", type);
-			e.addToBody("fault", "id");
-			e.addToBody("id", id);
-		}
-		return e;
 	}
 }
